[Optimize flamechart][1/4] Add flamechart to ReactProfilerData
This commit begins a stack of PRs that optimizes our flamechart.
Broadly, we need to replace the use of Speedscope's flamechart types
with our own types that are optimized for our rendering code.

As flamechart was always moved around together with ReactProfilerData,
it makes sense to move flamechart into it. That also necessitates moving
preprocessFlamechart into preprocessData.

This PR works towards #50.

* `yarn flow`: errors present, but no errors in affected code.
* `yarn lint`
* `yarn start`
