+++ src/App.js
@@ -1,9 +1,8 @@
 // @flow
 
-import type {FlamechartData, ReactProfilerData} from './types';
+import type {ReactProfilerData} from './types';
 
-import React, {useState, useCallback} from 'react';
-import {unstable_batchedUpdates} from 'react-dom';
+import React, {useState} from 'react';
 
 import ImportPage from './ImportPage';
 import CanvasPage from './CanvasPage';
@@ -12,22 +11,10 @@ export default function App() {
   const [profilerData, setProfilerData] = useState<ReactProfilerData | null>(
     null,
   );
-  const [flamechart, setFlamechart] = useState<FlamechartData | null>(null);
 
-  const handleDataImported = useCallback(
-    (
-      importedProfilerData: ReactProfilerData,
-      importedFlamechart: FlamechartData,
-    ) => {
-      unstable_batchedUpdates(() => {
-        setProfilerData(importedProfilerData);
-        setFlamechart(importedFlamechart);
-      });
-    },
-  );
-  if (profilerData && flamechart) {
-    return <CanvasPage profilerData={profilerData} flamechart={flamechart} />;
+  if (profilerData) {
+    return <CanvasPage profilerData={profilerData} />;
   } else {
-    return <ImportPage onDataImported={handleDataImported} />;
+    return <ImportPage onDataImported={setProfilerData} />;
   }
 }
